Add files via upload
FoxPro files are needed to process the .txt files from the tide gauge to convert them to .XLS and .csv files with each row giving depth or mean depth. The -5 and -6.xls are 5 and 6 minute average depths. The .xls and .csv are files showing every depth measured i.e. a row every 2 seconds. The reason for the .csv files is because the .xls files are restricted to 32767 records.